requirements.txt
# Core dependencies
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-cloud-secret-manager==2.16.4
google-api-python-client==2.108.0

# OpenAI
openai==1.3.0

# Social Media APIs
tweepy==4.14.0

# HTTP and utilities
requests==2.31.0
python-dotenv==1.0.0

# Optional but recommended
Pillow==10.1.0  # For image processing if needed
moviepy==1.0.3  # For video processing if needed
